+++ packages/fluentui/react-northstar/src/themes/teams/componentStyles.ts
@@ -97,6 +97,8 @@ export { skeletonLineStyles as SkeletonLine } from './components/Skeleton/skelet
 export { skeletonShapeStyles as SkeletonShape } from './components/Skeleton/skeletonShapeStyles';
 export { skeletonButtonStyles as SkeletonButton } from './components/Skeleton/skeletonButtonStyles';
 export { skeletonTextStyles as SkeletonText } from './components/Skeleton/skeletonTextStyles';
+export { skeletonInputStyles as SkeletonInput } from './components/Skeleton/skeletonInputStyles';
+export { skeletonAvatarStyles as SkeletonAvatar } from './components/Skeleton/skeletonAvatarStyles';
 
 export { sliderStyles as Slider } from './components/Slider/sliderStyles';