+++ packages/fluentui/react-northstar/src/themes/teams/components/Skeleton/skeletonVariables.ts
@@ -6,6 +6,7 @@ export interface SkeletonVariables {
   lineBackground?: string;
   shapeBackground?: string;
   animationBackground?: string;
+  animationBackgroundSecondary?: string;
 
   // Skeleton Button
   buttonHeight?: string;
@@ -23,6 +24,21 @@ export interface SkeletonVariables {
   textMediumHeight?: string;
   textLargeHeight?: string;
   textLargerHeight?: string;
+
+  // Skeleton Input
+  inputHeight?: string;
+  inputWidth?: string;
+  inputBackground?: string;
+
+  // Skeleton Avatar
+  avatarBackground?: string;
+  avatarSmallest?: string;
+  avatarSmaller?: string;
+  avatarSmall?: string;
+  avatarMedium?: string;
+  avatarLarge?: string;
+  avatarLarger?: string;
+  avatarLargest?: string;
 }
 
 export const skeletonVariables = (siteVariables): SkeletonVariables => {
@@ -30,6 +46,7 @@ export const skeletonVariables = (siteVariables): SkeletonVariables => {
     lineBackground: siteVariables.colorScheme.default.background4,
     shapeBackground: siteVariables.colorScheme.default.background4,
     animationBackground: siteVariables.colorScheme.default.background,
+    animationBackgroundSecondary: siteVariables.colorScheme.default.background5,
     lineMargin: `0 0 ${pxToRem(2)} 0`,
     shapeMargin: `0 0 ${pxToRem(2)} 0`,
 
@@ -49,5 +66,20 @@ export const skeletonVariables = (siteVariables): SkeletonVariables => {
     textMediumHeight: pxToRem(19),
     textLargeHeight: pxToRem(24),
     textLargerHeight: pxToRem(32),
+
+    // Skeleton Input
+    inputHeight: pxToRem(32),
+    inputWidth: pxToRem(154),
+    inputBackground: siteVariables.colorScheme.default.background4,
+
+    // Skeleton Avatar
+    avatarBackground: siteVariables.colorScheme.default.background4,
+    avatarSmallest: pxToRem(20),
+    avatarSmaller: pxToRem(24),
+    avatarSmall: pxToRem(28),
+    avatarMedium: pxToRem(32),
+    avatarLarge: pxToRem(44),
+    avatarLarger: pxToRem(64),
+    avatarLargest: pxToRem(96),
   };
 };